+++ src/containers/member/join/JoinForm.tsx
@@ -2,16 +2,16 @@
 
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
-import FirstBtnArea from '@/containers/(member)/join/FirstBtnArea'
-import IdInput from '@/containers/(member)/join/IdInput'
-import NicknameInput from '@/containers/(member)/join/NicknameInput'
-import PhoneVerification from '@/containers/(member)/join/PhoneVerification'
-import ProfileImgArea from '@/containers/(member)/join/ProfileImgArea'
-import Pw2Input from '@/containers/(member)/join/Pw2Input'
-import PwInput from '@/containers/(member)/join/PwInput'
-import SecondBtnArea from '@/containers/(member)/join/SecondBtnArea'
-import SelectCategory from '@/containers/(member)/join/SelectCategory'
-import { useFirstStore, useSecondStore } from '@/containers/(member)/join/store'
+import FirstBtnArea from '@/containers/member/join/FirstBtnArea'
+import IdInput from '@/containers/member/join/IdInput'
+import NicknameInput from '@/containers/member/join/NicknameInput'
+import PhoneVerification from '@/containers/member/join/PhoneVerification'
+import ProfileImgArea from '@/containers/member/join/ProfileImgArea'
+import Pw2Input from '@/containers/member/join/Pw2Input'
+import PwInput from '@/containers/member/join/PwInput'
+import SecondBtnArea from '@/containers/member/join/SecondBtnArea'
+import SelectCategory from '@/containers/member/join/SelectCategory'
+import { useFirstStore, useSecondStore } from '@/containers/member/join/store'
 import DuckOne from '@/public/svgs/duck/duckOne.svg'
 import { join } from '@/utils/memberApi'
 import { uploadImage } from '@/utils/uploadImage'
