+++ src/containers/member/join/NicknameInput.tsx
@@ -1,5 +1,5 @@
 import { TiDelete } from 'react-icons/ti'
-import { useFirstStore } from '@/containers/(member)/join/store'
+import { useFirstStore } from '@/containers/member/join/store'
 import { duplicationCheckNick } from '@/utils/memberApi'
 
 export default function NicknameInput() {
