+++ src/containers/member/join/SecondBtnArea.tsx
